rename func for more intend revealing
